+++ src/main/java/roomescape/controller/TimeController.java
@@ -1,0 +1,28 @@
+package roomescape.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.dto.response.ReservationTimeResponse;
+import roomescape.service.ReservationTimeQueryService;
+
+import java.time.LocalDate;
+import java.util.List;
+
+@RestController
+@RequestMapping("/times")
+@RequiredArgsConstructor
+public class TimeController {
+
+    private final ReservationTimeQueryService reservationTimeQueryService;
+
+    @GetMapping
+    public ResponseEntity<List<ReservationTimeResponse>> getAvailableTimes(
+            @RequestParam("date") LocalDate date,
+            @RequestParam("themeId") long themeId) {
+        return ResponseEntity.ok(reservationTimeQueryService.findAvailableReservationTimes(date, themeId));
+    }
+}
